Revert "Thread manager instantiation in the validator (#4603)"
This reverts commit e48672c122baf62e936d4d489d20ff7ad7ba53f9.
