+++ core/Cargo.toml
@@ -15,7 +15,6 @@ codecov = { repository = "solana-labs/solana", branch = "master", service = "git
 
 [dependencies]
 agave-banking-stage-ingress-types = { workspace = true }
-agave-thread-manager = { workspace = true }
 agave-transaction-view = { workspace = true }
 ahash = { workspace = true }
 anyhow = { workspace = true }
