+++ local-cluster/src/validator_configs.rs
@@ -73,7 +73,6 @@ pub fn safe_clone_config(config: &ValidatorConfig) -> ValidatorConfig {
         replay_forks_threads: config.replay_forks_threads,
         replay_transactions_threads: config.replay_transactions_threads,
         tvu_shred_sigverify_threads: config.tvu_shred_sigverify_threads,
-        thread_manager_config: config.thread_manager_config.clone(),
         delay_leader_block_for_pending_fork: config.delay_leader_block_for_pending_fork,
     }
 }
