--- thread-manager/src/default_config.toml
@@ -1,1 +1,0 @@
-# This is deliberately left empty to populate as integration process proceeds.
